+++ R/gllvm.TMB.R
@@ -354,7 +354,11 @@ gllvm.TMB <- function(y, X = NULL, lv.X = NULL, xr = matrix(0), formula = NULL, 
         }
         if(randomB!=FALSE & randomB!="iid"){
           sigmab_lv <- fit$sigmab_lv
-          if(ncol(csBlv)==2)sigmab_lv <- c(sigmab_lv, fit$sigmab_cors[csBlv])
+          if(ncol(csBlv)==2 && starting.val == "res"){
+            sigmab_lv <- c(sigmab_lv, fit$sigmab_cors[csBlv])
+          }else{
+            sigmab_lv <- c(sigmab_lv, rep(1e-3, nrow(csBlv)))
+          }
         }
         if(num.lv>1&(num.lv.c+num.RR)==0){
           lambdas[upper.tri(lambdas)] <- 0